requirements.txt
alabaster==0.7.7
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-appconf==1.0.1
django-debug-toolbar==1.2.1
django-debug-toolbar-mongo==0.1.10
django-extensions==1.6.1
django-toolbelt==0.0.1
docutils==0.12
gnureadline==6.3.3
gunicorn==19.3.0
ipython==2.1.0
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
mongoengine==0.8.7
nose==1.3.7
numpydoc==0.5
Pillow==3.0.0
pinax-blog==4.3.0
psycopg2==2.6
Pygments==2.0.2
#pymongo==3.2
pymongo==2.8.1
python-slugify==1.1.4
pytz==2015.7
pyzmq==15.1.0
requests==2.2.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
sqlparse==0.1.16
static3==0.6.1
tornado==4.3
unbabel-py==0.25
Unidecode==0.4.18
whitenoise==3.2.1
waitress==0.8.9
